header file consolidation
