+++ src/AP_ID3v2_tags.cpp
@@ -22,9 +22,6 @@
 #include "AtomicParsley.h"
 #include "AP_commons.h"
 #include "AP_ID3v2_tags.h"
-#include "APar_zlib.h"
-#include "AP_iconv.h"
-#include "APar_uuid.h"
 #include "AP_CDTOC.h"
 #include "AP_arrays.h"
