+++ src/AP_MetadataListings.cpp
@@ -21,10 +21,8 @@
 
 #include "AtomicParsley.h"
 #include "AP_commons.h"
-#include "AP_iconv.h"
 #include "AP_ID3v2_tags.h"
 #include "AP_arrays.h"
-#include "APar_uuid.h"
 
 #include "AP_MetadataListings.h"
