+++ src/AP_commons.cpp
@@ -27,7 +27,6 @@
 
 #include "AtomicParsley.h"
 #include "AP_commons.h"
-#include "AP_iconv.h"
 
 
 ///////////////////////////////////////////////////////////////////////////////////////
