+++ src/APar_sha1.cpp
@@ -32,7 +32,6 @@
 */
 
 #include "AtomicParsley.h"
-#include "APar_sha1.h"
 
 /* SWAP does an endian swap on architectures that are little-endian,
    as SHA1 needs some data in a big-endian form.  */
