+++ src/main.cpp
@@ -29,9 +29,7 @@
 
 #include "AP_commons.h"
 #include "AP_AtomExtracts.h"
-#include "AP_iconv.h"                 /* for xmlInitEndianDetection used in endian utf16 conversion */
 #include "AP_arrays.h"
-#include "APar_uuid.h"
 #include "AP_ID3v2_tags.h"
 #include "AP_MetadataListings.h"
